requirements.txt
dj-database-url==0.5.0
Django==2.2.4
django-crispy-forms==1.12.0
django-pygments==0.4
django-pygments-renderer==0.0.1
django-widget-tweaks==1.4.8
Pillow==8.3.1
psycopg2==2.9.1
Pygments==2.4.2
python-decouple==3.4
pytz==2019.2
sqlparse==0.3.0
whitenoise==5.2.0
